requirements.txt
aiosmtplib==2.0.2
annotated-types==0.5.0
anyio==3.7.1
atlastk==0.13.3
blinker==1.6.2
dnspython==2.4.2
email-validator==2.0.0.post2
fastapi==0.103.1
fastapi-mail==1.4.1
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
python-decouple==3.8
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
tortoise==0.1.1
typing_extensions==4.7.1
